Server/requirements.txt
absl-py==2.0.0
ai4bharat-transliteration==1.1.3
alabaster==0.7.13
antlr4-python3-runtime==4.8
appnope==0.1.3
asttokens==2.4.1
astunparse==1.6.3
Babel==2.13.1
bitarray==2.8.3
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.0
contourpy==1.2.0
cycler==0.12.1
Cython==3.0.6
debugpy==1.8.0
decorator==5.1.1
dm-tree==0.1.8
docutils==0.20.1
exceptiongroup==1.2.0
executing==2.0.1
fairseq==0.12.2
filelock==3.13.1
Flask==3.0.0
Flask-Cors==4.0.0
flatbuffers==23.5.26
fonttools==4.46.0
fsspec==2023.10.0
gast==0.5.4
gevent==23.9.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-pasta==0.2.0
greenlet==3.0.1
grpcio==1.59.3
h5py==3.10.0
huggingface-hub==0.19.4
hydra-core==1.0.7
idna==3.6
imagesize==1.4.1
importlib-metadata==6.8.0
importlib-resources==6.1.1
indic-nlp-library==0.92
ipykernel==6.27.1
ipython==8.18.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.5.0
keras==2.15.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
libclang==16.0.6
lxml==4.9.3
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.2.0
mock==5.1.0
Morfessor==2.0.6
mpmath==1.3.0
namex==0.0.7
nest-asyncio==1.5.8
networkx==3.2.1
nltk==3.8.1
numpy==1.26.2
oauthlib==3.2.2
omegaconf==2.0.6
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.3
parso==0.8.3
pexpect==4.9.0
Pillow==10.1.0
platformdirs==4.0.0
portalocker==2.8.2
progressbar2==4.2.0
prompt-toolkit==3.0.41
protobuf==4.23.4
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydload==1.0.9
Pygments==2.17.2
pyparsing==3.1.1
python-crfsuite==0.9.9
python-dateutil==2.8.2
python-utils==3.8.1
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rsa==4.9
sacrebleu==2.3.3
sacremoses==0.1.1
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
sklearn-crfsuite==0.3.6
snowballstemmer==2.2.0
Sphinx==7.2.6
sphinx-argparse==0.4.0
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-macos==2.15.0
termcolor==2.3.0
threadpoolctl==3.2.0
tokenizers==0.15.0
torch==2.1.1
torchaudio==2.1.1
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
transformers==4.35.2
typing_extensions==4.8.0
tzdata==2023.3
ujson==5.8.0
urduhack==0.1.4
urllib3==2.1.0
wcwidth==0.2.12
Werkzeug==3.0.1
wrapt==1.14.1
zipp==3.17.0
zope.event==5.0
zope.interface==6.1
